SharePoint error codes need to be Long, as they must not be negative and may exceed 0x80000000
git-svn-id: http://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@31717 7033f55c-6c29-4ad8-afba-765cb616e738

+++ root/modules/sharepoint/amp/source/java/org/alfresco/module/vti/handler/VtiHandlerException.java
@@ -61,7 +61,7 @@ public class VtiHandlerException extends RuntimeException
     /**
      * What SharePoint error code to return
      */
-    private int errorCode = -1;
+    private long errorCode = -1;
     
     /**
      * What Error this is, from the Enumeration List
@@ -139,7 +139,7 @@ public class VtiHandlerException extends RuntimeException
     /**
      * Return the error code, or -1 if not known
      */
-    public int getErrorCode()
+    public long getErrorCode()
     {
         return errorCode;
     }
